+++ modules/core/greetd.nix
@@ -9,7 +9,22 @@
     settings = {
       default_session = {
         user = username;
-        command = "${pkgs.greetd.tuigreet}/bin/tuigreet --time --cmd Hyprland"; # start Hyprland with a TUI login manager
+        command = ''
+          ${pkgs.greetd.tuigreet}/bin/tuigreet \
+            --time \
+            --asterisks \
+            --user-menu \
+            --remember \
+            --remember-session \
+            --cmd Hyprland \
+            --width 40 \
+            --greeting "Welcome to DuhzitOS" \
+            --text-color "#FFFFFF" \
+            --text-highlight-color "#FDCA40" \
+            --prompt-color "#66C3FF" \
+            --prompt-highlight-color "#F64740" \
+            --border-color "#550C18"
+        '';
       };
     };
   };